+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # WaterDrop changelog
 
 ## 2.8.3 (Unreleased)
+- [Enhancement] Support producing messages with arrays of strings in headers (KIP-82).
 - [Refactor] Introduce a `bin/verify_topics_naming` script to ensure proper test topics naming convention.
 - [Refactor] Remove factory bot and active support requirement in tests/dev.